fix: Circular dependencies
