+++ api/src/utils/types.ts
@@ -1,5 +1,6 @@
 import { stringCamelCase } from '@polkadot/util';
 import { REGULAR_EXP } from './regexp';
+import { CreateTypeError } from '../errors';
 
 export function transformTypes(types: object): { [key: string]: object } {
   return Object.values(types).reduce((res, types): object => ({ ...res, ...types }), {});
@@ -17,4 +18,15 @@ export function typeIsGeneric(type: string) {
   } else {
     return false;
   }
+}
+
+export function typeIsString(type: string): boolean {
+  return ['string', 'utf8', 'utf-8', 'text'].includes(type.toLowerCase());
+}
+
+export function checkTypeAndPayload(type: string, payload: unknown): string {
+  if (payload === undefined) {
+    throw new CreateTypeError('Payload is not specified');
+  }
+  return type || 'Bytes';
 }
